requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
dnspython==2.7.0
email_validator==2.2.0
greenlet==3.2.1
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
phonenumbers==9.0.0
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
typing-inspection==0.4.0
typing_extensions==4.12.2
cryptography>=39.0.0
